copy code in pyopencv_rotationwarper.hpp
